refactor: useTreeMap의 marker 관련 함수를 useMarMarker로 이동
