+++ frontend/src/hooks/TreeMap/useMapMarkers.tsx
@@ -1,7 +1,8 @@
-import { useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import { Root, createRoot } from 'react-dom/client';
 import ClusterOverlay from '@/components/Tree/ClusterOverlay';
 import useTreeMap from '@/hooks/TreeMap/useTreeMap';
+import treeImage from '@/assets/TREE_01.png';
 
 interface Tree {
   id: number;
@@ -21,24 +22,26 @@ interface UseMapMarkersParams {
   isClusterView: boolean;
   trees: Tree[];
   treeClusters: TreeCluster[];
-  addCustomOverlay: ReturnType<typeof useTreeMap>['addCustomOverlay'];
-  clearMarkers: ReturnType<typeof useTreeMap>['clearMarkers'];
-  addMarker: ReturnType<typeof useTreeMap>['addMarker'];
   onMarkerClick: (treeId: number) => void;
   onClusterClick: ({ latitude, longitude }: { latitude: number; longitude: number }) => void;
 }
+
+const { kakao } = window;
+
+const MARKER_IMAGE: Record<string, string> = {
+  TREE_01: treeImage,
+};
 
 const useMapMarkers = ({
   map,
   isClusterView,
   trees,
   treeClusters,
-  addCustomOverlay,
-  clearMarkers,
-  addMarker,
   onMarkerClick,
   onClusterClick,
 }: UseMapMarkersParams) => {
+  const markersRef = useRef<(typeof kakao.maps.Marker | typeof kakao.maps.CustomOverlay)[]>([]);
+
   useEffect(() => {
     if (!map) return;
 
@@ -73,6 +76,68 @@ const useMapMarkers = ({
       roots.forEach((root) => root.unmount());
     };
   }, [map, trees, treeClusters, isClusterView]);
+
+  const addCustomOverlay = (
+    targetMap: typeof kakao.maps.Map,
+    latitude: number,
+    longitude: number,
+    content: HTMLElement,
+    onClick?: () => void,
+  ) => {
+    if (!targetMap) return;
+
+    const position = new kakao.maps.LatLng(latitude, longitude);
+
+    const overlay = new kakao.maps.CustomOverlay({
+      position,
+      content,
+      clickable: true,
+    });
+
+    if (onClick) {
+      kakao.maps.event.addListener(overlay, 'click', onClick);
+    }
+
+    overlay.setMap(targetMap);
+    markersRef.current.push(overlay);
+  };
+
+  const createMarkerImage = (imageCode: string) => {
+    return new kakao.maps.MarkerImage(MARKER_IMAGE[imageCode], new kakao.maps.Size(50, 55), {
+      offset: new kakao.maps.Point(25, 55),
+    });
+  };
+
+  const addMarker = (
+    targetMap: typeof kakao.maps.Map,
+    latitude: number,
+    longitude: number,
+    imageCode: string,
+    onClick?: () => void,
+  ) => {
+    if (!targetMap) return;
+
+    const position = new kakao.maps.LatLng(latitude, longitude);
+    const image = createMarkerImage(imageCode);
+
+    const marker = new kakao.maps.Marker({
+      position,
+      image,
+      clickable: true,
+    });
+
+    if (onClick) {
+      kakao.maps.event.addListener(marker, 'click', onClick);
+    }
+
+    marker.setMap(targetMap);
+    markersRef.current.push(marker);
+  };
+
+  const clearMarkers = () => {
+    markersRef.current.forEach((marker) => marker.setMap(null));
+    markersRef.current = [];
+  };
 };
 
 export default useMapMarkers;
