+++ frontend/src/hooks/TreeMap/useTreeMap.ts
@@ -1,14 +1,9 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { DEFAULT_LATITUDE, DEFAULT_LONGITUDE } from '@/constants/map';
-import treeImage from '@/assets/TREE_01.png';
 
 const { kakao } = window;
 const DEFAULT_ZOOM_LEVEL = 3;
 const USER_LOCATION_STORAGE_KEY = 'userLocation';
-
-const MARKER_IMAGE: Record<string, string> = {
-  TREE_01: treeImage,
-};
 
 interface Coordinates {
   latitude: number;
@@ -54,19 +49,12 @@ const calculateMapBounds = (map: typeof kakao.maps.Map | null): Bounds => {
   };
 };
 
-const createMarkerImage = (imageCode: string) => {
-  return new kakao.maps.MarkerImage(MARKER_IMAGE[imageCode], new kakao.maps.Size(50, 55), {
-    offset: new kakao.maps.Point(25, 55),
-  });
-};
-
 const useTreeMap = () => {
   const mapRef = useRef<HTMLDivElement | null>(null);
   const [map, setMap] = useState<typeof kakao.maps.Map | null>(null);
   const [zoom, setZoom] = useState<number>(DEFAULT_ZOOM_LEVEL);
   const [centerPosition, setCenterPosition] = useState<Coordinates>(locationStorage.get);
   const [bounds, setBounds] = useState<Bounds>(() => calculateMapBounds(null));
-  const markersRef = useRef<(typeof kakao.maps.Marker | typeof kakao.maps.CustomOverlay)[]>([]);
 
   useEffect(() => {
     if (!mapRef.current || !kakao?.maps) return;
@@ -91,68 +79,6 @@ const useTreeMap = () => {
     return () => {
       kakao.maps.event.removeListener(mapInstance, 'zoom_changed', handleZoomChange);
     };
-  }, []);
-
-  const addMarker = useCallback(
-    (
-      targetMap: typeof kakao.maps.Map,
-      latitude: number,
-      longitude: number,
-      imageCode: string,
-      onClick?: () => void,
-    ) => {
-      if (!targetMap) return;
-
-      const position = new kakao.maps.LatLng(latitude, longitude);
-      const image = createMarkerImage(imageCode);
-
-      const marker = new kakao.maps.Marker({
-        position,
-        image,
-        clickable: true,
-      });
-
-      if (onClick) {
-        kakao.maps.event.addListener(marker, 'click', onClick);
-      }
-
-      marker.setMap(targetMap);
-      markersRef.current.push(marker);
-    },
-    [],
-  );
-
-  const addCustomOverlay = useCallback(
-    (
-      targetMap: typeof kakao.maps.Map,
-      latitude: number,
-      longitude: number,
-      content: HTMLElement,
-      onClick?: () => void,
-    ) => {
-      if (!targetMap) return;
-
-      const position = new kakao.maps.LatLng(latitude, longitude);
-
-      const overlay = new kakao.maps.CustomOverlay({
-        position,
-        content,
-        clickable: true,
-      });
-
-      if (onClick) {
-        kakao.maps.event.addListener(overlay, 'click', onClick);
-      }
-
-      overlay.setMap(targetMap);
-      markersRef.current.push(overlay);
-    },
-    [],
-  );
-
-  const clearMarkers = useCallback(() => {
-    markersRef.current.forEach((marker) => marker.setMap(null));
-    markersRef.current = [];
   }, []);
 
   const updateCenterPosition = () => {
@@ -198,9 +124,6 @@ const useTreeMap = () => {
     centerPosition,
     bounds,
     zoom,
-    addMarker,
-    addCustomOverlay,
-    clearMarkers,
     updateCenterPosition,
     updateBounds,
     updatePosition,
